\docType{data}
\name{BYB1_G07_pruned}
\alias{BYB1_G07_pruned}
\title{BYB1_G07_pruned}
\format{A \code{data.frame} with 12 genes (rows) measured at 11 time points (columns)}
\description{
A real data set taken from Lang et al, 2013
(PMID:23873039).  See the accompanying vignette for more
details.
}
\author{
Alex Murison \email{Alexander.Murison@icr.ac.uk} and
Christopher Wardell \email{Christopher.Wardell@icr.ac.uk}
}
\seealso{
\code{\link{ga-class}}, \code{\link{ga}},
\code{\link{gauchoReport}},
\code{\link{gaucho_simple_data}},
\code{\link{gaucho_hidden_data}},
\code{\link{gaucho_synth_data}},
\code{\link{gaucho_synth_data_jittered}},
\code{\link{BYB1_G07_pruned}}
}

