\docType{data}
\name{gaucho_hidden_data}
\alias{gaucho_hidden_data}
\title{gaucho_hidden_data}
\format{A \code{data.frame} with 5 mutations (rows) measured at 3 time points (columns)}
\description{
A synthetic data set provided to illustrate how to use
gaucho.  See the accompanying vignette for more details.
}
\author{
Alex Murison \email{Alexander.Murison@icr.ac.uk} and
Christopher Wardell \email{Christopher.Wardell@icr.ac.uk}
}
\seealso{
\code{\link{ga-class}}, \code{\link{ga}},
\code{\link{gauchoReport}},
\code{\link{gaucho_simple_data}},
\code{\link{gaucho_hidden_data}},
\code{\link{gaucho_synth_data}},
\code{\link{gaucho_synth_data_jittered}},
\code{\link{BYB1_G07_pruned}}
}

