\docType{data}
\name{gaucho_synth_data_jittered}
\alias{gaucho_synth_data_jittered}
\title{gaucho_synth_data_jittered}
\format{A \code{data.frame} with 90 genes (rows) measured at 4 time points (columns), with added noise.}
\description{
A synth data set provided to illustrate how to use gaucho.
See the accompanying vignette for more details.
}
\author{
Alex Murison \email{Alexander.Murison@icr.ac.uk} and
Christopher Wardell \email{Christopher.Wardell@icr.ac.uk}
}
\seealso{
\code{\link{ga-class}}, \code{\link{ga}},
\code{\link{gauchoReport}},
\code{\link{gaucho_simple_data}},
\code{\link{gaucho_hidden_data}},
\code{\link{gaucho_synth_data}},
\code{\link{gaucho_synth_data_jittered}},
\code{\link{BYB1_G07_pruned}}
}

